geotiff tests: consolidate metadata / CRS-write / COG-BigTIFF / reader tails (#2439)
Cluster 15 of long-tail epic #2424. Tests-only. Folds 15 top-level
test files into the existing unit/, read/, and write/ homes via four
small sub-clusters:

* Metadata (5): new `unit/test_metadata.py` plus extending
  `unit/test_safe_xml.py` for the safe extra_tags filter (#1657).
* CRS write (6): new `write/test_crs.py`. (`read/test_crs.py` from
  #2390 PR 3 keeps the read side.)
* COG / BigTIFF tail (2): extends `write/test_bigtiff.py`.
* Reader tail (2): extends `read/test_basic.py`.

Each consolidated section is suffixed with the source issue number so
helpers and class names stay collision-free across the merged
sections. Per-file `_gpu_only` probes collapse onto the shared
`requires_gpu` marker from `_helpers/markers.py`.

`docs/source/reference/release_gate_geotiff.rst`, `_attrs.py`, and
the prose pointers in `test_round_trip_invariants.py`,
`test_roundtrip_properties.py`, `test_crs_fail_closed_gpu_1929.py`,
and `test_remaining_fail_closed_1987.py` are updated to point at the
new paths so the release-gate checklist-parity test stays green.

`CLUSTER_AUDIT_TAILS.md` maps every old `file::test` to its new
location; the audit file is deleted on a final commit before merge
per the epic's hard gate.

File count drops by 13 (15 deleted, 2 added).

Verification (on this checkout):

- `pytest xrspatial/geotiff/tests/unit/ xrspatial/geotiff/tests/read/
  xrspatial/geotiff/tests/write/`: 2719 passed, 6 skipped.
- `pytest xrspatial/geotiff/tests/ --ignore=integration`: 5346
  passed, 68 skipped, 2 xfailed.
- `pytest .../release_gates/test_stable_features.py
  ::test_release_gate_cites_only_existing_test_files`: 1 passed.
