+++ xrspatial/geotiff/tests/read/test_basic.py
@@ -110,3 +110,298 @@ class TestBandValidationBackendParity:
 
         assert "band=3 out of range" in str(eager_exc.value)
         assert "band=3 out of range" in str(dask_exc.value)
+
+
+# =============================================================================
+# Section: open_geotiff missing_sources (#1810)
+# =============================================================================
+#
+# Original: ``test_open_geotiff_missing_sources_1810.py``.
+#
+# ``open_geotiff`` did not accept ``missing_sources`` and did not
+# forward it to ``read_vrt`` when the source was a VRT. The
+# api-consistency sweep on 2026-05-13 flagged that ``read_vrt`` had
+# gained a ``missing_sources='warn'|'raise'`` policy kwarg (#1806) but
+# the documented dispatcher entry point ``open_geotiff`` did not
+# expose it. Same class of dispatcher-drops-backend-kwarg bug as
+# #1561, #1605, #1685, #1795.
+import inspect  # noqa: E402
+
+from xrspatial.geotiff import GeoTIFFFallbackWarning, open_geotiff, to_geotiff  # noqa: E402
+from xrspatial.geotiff import write_vrt as _write_vrt_1810  # noqa: E402
+
+
+def _write_missing_source_vrt_1810(path):
+    """Write a VRT pointing at a non-existent source file."""
+    path.write_text(
+        '<VRTDataset rasterXSize="2" rasterYSize="2">\n'
+        '  <VRTRasterBand dataType="Byte" band="1">\n'
+        '    <SimpleSource>\n'
+        '      <SourceFilename relativeToVRT="1">missing.tif'
+        '</SourceFilename>\n'
+        '      <SourceBand>1</SourceBand>\n'
+        '      <SrcRect xOff="0" yOff="0" xSize="2" ySize="2"/>\n'
+        '      <DstRect xOff="0" yOff="0" xSize="2" ySize="2"/>\n'
+        '    </SimpleSource>\n'
+        '  </VRTRasterBand>\n'
+        '</VRTDataset>\n'
+    )
+
+
+def test_open_geotiff_accepts_missing_sources_1810():
+    """The signature exposes ``missing_sources`` so IDE autocomplete and
+    ``inspect.signature`` see the kwarg without parsing the docstring.
+    """
+    sig = inspect.signature(open_geotiff)
+    assert 'missing_sources' in sig.parameters
+
+
+def test_open_geotiff_vrt_missing_sources_warn_1810(tmp_path):
+    vrt = tmp_path / "tmp_1810_missing_warn.vrt"
+    _write_missing_source_vrt_1810(vrt)
+
+    with pytest.warns(GeoTIFFFallbackWarning, match="could not be read"):
+        da = open_geotiff(str(vrt), missing_sources='warn')
+
+    assert 'vrt_holes' in da.attrs
+    assert da.attrs['vrt_holes'][0]['source'].endswith('missing.tif')
+
+
+def test_open_geotiff_vrt_missing_sources_raise_1810(tmp_path):
+    vrt = tmp_path / "tmp_1810_missing_raise.vrt"
+    _write_missing_source_vrt_1810(vrt)
+
+    with pytest.raises((OSError, ValueError)):
+        open_geotiff(str(vrt), missing_sources='raise')
+
+
+def test_open_geotiff_vrt_missing_sources_validates_policy_1810(tmp_path):
+    vrt = tmp_path / "tmp_1810_missing_bad_policy.vrt"
+    _write_missing_source_vrt_1810(vrt)
+
+    with pytest.raises(ValueError, match="missing_sources"):
+        open_geotiff(str(vrt), missing_sources='ignore')
+
+
+def test_open_geotiff_rejects_missing_sources_on_tif_1810(tmp_path):
+    arr = np.arange(16, dtype=np.uint16).reshape(4, 4)
+    da = xr.DataArray(
+        arr, dims=['y', 'x'],
+        coords={'y': np.arange(4, dtype=np.float64),
+                'x': np.arange(4, dtype=np.float64)},
+        attrs={'crs': 4326},
+    )
+    tif_path = tmp_path / "single_tile.tif"
+    to_geotiff(da, str(tif_path))
+
+    with pytest.raises(ValueError, match="missing_sources only applies"):
+        open_geotiff(str(tif_path), missing_sources='raise')
+
+
+def test_open_geotiff_vrt_without_missing_sources_kwarg_still_works_1810(tmp_path):
+    arr_a = np.arange(16, dtype=np.uint16).reshape(4, 4)
+    da_a = xr.DataArray(
+        arr_a, dims=['y', 'x'],
+        coords={'y': np.array([0.5, 1.5, 2.5, 3.5]),
+                'x': np.array([0.5, 1.5, 2.5, 3.5])},
+        attrs={'crs': 4326},
+    )
+    tile_a = tmp_path / "tile_a.tif"
+    to_geotiff(da_a, str(tile_a))
+
+    arr_b = np.arange(16, 32, dtype=np.uint16).reshape(4, 4)
+    da_b = xr.DataArray(
+        arr_b, dims=['y', 'x'],
+        coords={'y': np.array([0.5, 1.5, 2.5, 3.5]),
+                'x': np.array([4.5, 5.5, 6.5, 7.5])},
+        attrs={'crs': 4326},
+    )
+    tile_b = tmp_path / "tile_b.tif"
+    to_geotiff(da_b, str(tile_b))
+
+    vrt_path = tmp_path / "mosaic_1810.vrt"
+    _write_vrt_1810(str(vrt_path), [str(tile_a), str(tile_b)])
+
+    da = open_geotiff(str(vrt_path))
+    assert da.shape == (4, 8)
+
+
+# =============================================================================
+# Section: Reader strips/tiles/array (low-level) and partial-tile validation
+# =============================================================================
+#
+# Original: ``test_reader.py`` (general low-level reader coverage) and
+# the ``#1486`` partial-tile validation block within it.
+#
+# Covers ``_read_strips`` / ``_read_tiles`` / ``read_to_array`` happy
+# paths and the truncated-tile rejection contract that turns opaque
+# numpy reshape errors into clear "size mismatch" diagnostics.
+from xrspatial.geotiff._dtypes import tiff_dtype_to_numpy  # noqa: E402
+from xrspatial.geotiff._header import parse_all_ifds, parse_header  # noqa: E402
+from xrspatial.geotiff._reader import _read_strips, _read_tiles, read_to_array  # noqa: E402
+
+from ..conftest import make_minimal_tiff  # noqa: E402
+
+
+class TestReadStrips_reader:
+    def test_float32_sequential(self):
+        """Read a simple float32 stripped TIFF and verify pixel values."""
+        expected = np.arange(16, dtype=np.float32).reshape(4, 4)
+        data = make_minimal_tiff(4, 4, np.dtype('float32'), pixel_data=expected)
+
+        header = parse_header(data)
+        ifds = parse_all_ifds(data, header)
+        ifd = ifds[0]
+        dtype = tiff_dtype_to_numpy(ifd.bits_per_sample, ifd.sample_format)
+
+        arr = _read_strips(data, ifd, header, dtype)
+        np.testing.assert_array_equal(arr, expected)
+
+    def test_uint16(self):
+        expected = np.arange(20, dtype=np.uint16).reshape(4, 5)
+        data = make_minimal_tiff(5, 4, np.dtype('uint16'), pixel_data=expected)
+
+        header = parse_header(data)
+        ifds = parse_all_ifds(data, header)
+        ifd = ifds[0]
+        dtype = tiff_dtype_to_numpy(ifd.bits_per_sample, ifd.sample_format)
+
+        arr = _read_strips(data, ifd, header, dtype)
+        np.testing.assert_array_equal(arr, expected)
+
+    def test_windowed_read(self):
+        expected = np.arange(64, dtype=np.float32).reshape(8, 8)
+        data = make_minimal_tiff(8, 8, np.dtype('float32'), pixel_data=expected)
+
+        header = parse_header(data)
+        ifds = parse_all_ifds(data, header)
+        ifd = ifds[0]
+        dtype = tiff_dtype_to_numpy(ifd.bits_per_sample, ifd.sample_format)
+
+        window = (2, 3, 6, 7)  # rows 2-5, cols 3-6
+        arr = _read_strips(data, ifd, header, dtype, window=window)
+        np.testing.assert_array_equal(arr, expected[2:6, 3:7])
+
+
+class TestReadTiles_reader:
+    def test_tiled_float32(self):
+        expected = np.arange(64, dtype=np.float32).reshape(8, 8)
+        data = make_minimal_tiff(
+            8, 8, np.dtype('float32'),
+            pixel_data=expected,
+            tiled=True,
+            tile_size=4,
+        )
+
+        header = parse_header(data)
+        ifds = parse_all_ifds(data, header)
+        ifd = ifds[0]
+        dtype = tiff_dtype_to_numpy(ifd.bits_per_sample, ifd.sample_format)
+
+        arr = _read_tiles(data, ifd, header, dtype)
+        np.testing.assert_array_equal(arr, expected)
+
+    def test_tiled_windowed(self):
+        expected = np.arange(64, dtype=np.float32).reshape(8, 8)
+        data = make_minimal_tiff(
+            8, 8, np.dtype('float32'),
+            pixel_data=expected,
+            tiled=True,
+            tile_size=4,
+        )
+
+        header = parse_header(data)
+        ifds = parse_all_ifds(data, header)
+        ifd = ifds[0]
+        dtype = tiff_dtype_to_numpy(ifd.bits_per_sample, ifd.sample_format)
+
+        window = (1, 2, 5, 6)
+        arr = _read_tiles(data, ifd, header, dtype, window=window)
+        np.testing.assert_array_equal(arr, expected[1:5, 2:6])
+
+
+class TestReadToArray_reader:
+    def test_local_file(self, tmp_path):
+        expected = np.arange(16, dtype=np.float32).reshape(4, 4)
+        tiff_data = make_minimal_tiff(4, 4, np.dtype('float32'), pixel_data=expected)
+        path = str(tmp_path / 'test.tif')
+        with open(path, 'wb') as f:
+            f.write(tiff_data)
+
+        arr, geo_info = read_to_array(path)
+        np.testing.assert_array_equal(arr, expected)
+
+    def test_geo_info(self, tmp_path):
+        tiff_data = make_minimal_tiff(
+            4, 4, np.dtype('float32'),
+            geo_transform=(-120.0, 45.0, 0.001, -0.001),
+            epsg=4326,
+        )
+        path = str(tmp_path / 'geo_test.tif')
+        with open(path, 'wb') as f:
+            f.write(tiff_data)
+
+        arr, geo_info = read_to_array(path)
+        assert geo_info.crs_epsg == 4326
+        assert geo_info.transform.origin_x == pytest.approx(-120.0)
+
+
+class TestPartialTileValidation_1486:
+    """Issue #1486: corrupt tile/strip data should raise a clear error.
+
+    Without validation a truncated deflate stream causes numpy.reshape to
+    raise an opaque "cannot reshape array of size N" with no hint of which
+    tile is at fault.  These tests pin the new behaviour: a clear ValueError
+    naming the size mismatch.
+    """
+
+    def _zero_out_last_tile_1486(self, path):
+        """Replace the last tile's compressed bytes with zeros so deflate
+        decodes a short stream."""
+        with open(path, 'rb') as f:
+            data = bytearray(f.read())
+        header = parse_header(bytes(data))
+        ifds = parse_all_ifds(bytes(data), header)
+        ifd = ifds[0]
+        if ifd.tile_offsets is not None:
+            offsets = ifd.tile_offsets
+            counts = ifd.tile_byte_counts
+        else:
+            offsets = ifd.strip_offsets
+            counts = ifd.strip_byte_counts
+        last_off = offsets[-1]
+        last_count = counts[-1]
+        zero_stream = b'\x78\x9c\x03\x00\x00\x00\x00\x01'
+        padded = zero_stream + b'\x00' * max(0, last_count - len(zero_stream))
+        for i, b in enumerate(padded[:last_count]):
+            data[last_off + i] = b
+        with open(path, 'wb') as f:
+            f.write(bytes(data))
+
+    def test_truncated_tile_raises_clear_error(self, tmp_path):
+        from xrspatial.geotiff._writer import write
+
+        pixels = np.arange(256, dtype=np.float32).reshape(16, 16)
+        path = str(tmp_path / 'truncated_1486.tif')
+        write(pixels, path, compression='deflate', tiled=True, tile_size=8)
+
+        self._zero_out_last_tile_1486(path)
+
+        with pytest.raises(ValueError) as exc:
+            read_to_array(path)
+        msg = str(exc.value)
+        assert 'size mismatch' in msg
+        assert 'expected' in msg
+        assert 'truncated or corrupt' in msg
+
+    def test_valid_edge_tile_still_works(self, tmp_path):
+        """Edge tiles in a valid file decompress to full tile size; the
+        validation should not flag this as corrupt."""
+        from xrspatial.geotiff._writer import write
+
+        pixels = np.arange(81, dtype=np.float32).reshape(9, 9)
+        path = str(tmp_path / 'edge_tiles_1486.tif')
+        write(pixels, path, compression='deflate', tiled=True, tile_size=4)
+
+        arr, _ = read_to_array(path)
+        np.testing.assert_array_equal(arr, pixels)
